Fix robot to servo coordinate conversion

+++ source/motor/motorik/gangstil.h
@@ -113,7 +113,7 @@ Point2D* raise_to_default_position();
 
 #ifdef IS_UNIT_TEST
 Point2D rotate_point_by_angle(Point2D original, float angle);
-Point2D robot_to_ik_coords(Point2D original, int leg);
+Point2D robot_to_ik_coords(Point2D original, size_t leg);
 float absf(float a);
 #endif
